+++ doc/whats-new.rst
@@ -61,6 +61,9 @@ New Features
   :py:class:`~core.groupby.DataArrayGroupBy`, inspired by pandas'
   :py:meth:`~pandas.core.groupby.GroupBy.get_group`.
   By `Deepak Cherian <https://github.com/dcherian>`_.
+- Implement :py:meth:`DataArray.unique` to return a flattened NumPy array
+  of unique values in the data array; convenient for iteration (:pull:`5090`).
+  By `Andrew Huang <https://github.com/ahuang11>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~